# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HostNepal CMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List all tickets
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [Open, In Progress, Resolved, Closed, all]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    put:
      operationId: updateTicket
      tags: [tickets]
      summary: Update a ticket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete a ticket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Ticket:
      type: object
      required:
        - id
        - ticketId
        - clientId
        - subject
        - status
        - createdAt
      properties:
        id:
          type: string
        ticketId:
          type: string
        clientId:
          type: string
        clientName:
          type: ["string", "null"]
        subject:
          type: string
        status:
          type: string
          enum: [Open, In Progress, Resolved, Closed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketInput:
      type: object
      required: [clientId, subject, status]
      properties:
        clientId:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [Open, In Progress, Resolved, Closed]
        notes:
          type: string
    TicketUpdate:
      type: object
      properties:
        subject:
          type: string
        status:
          type: string
          enum: [Open, In Progress, Resolved, Closed]
        notes:
          type: string
    TicketListResponse:
      type: object
      required: [tickets, total, page, limit, totalPages]
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...